Rag_test/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# RAG & Vector DB
chromadb==0.4.18
ollama==0.1.6

# PDF Processing
PyPDF2==3.0.1
PyMuPDF>=1.24.0  # Use newer version with pre-built wheels for Windows
Pillow==10.1.0

# Web Search
brave-search==0.1.5  # Brave Search API
tavily-python==0.3.0  # Alternative: Tavily API
requests==2.31.0

# Utilities
python-dotenv==1.0.0

# Frontend (Streamlit)
streamlit==1.29.0

# Optional: Progress tracking
tqdm==4.66.1